// src/shared/types/player.ts
export interface Player {
  id: number;
  name: string;
  age: number;
  dorsal: number;
  position: string;
  nationality: string;
  photo: string;
  height: string;
  weight: string;
  goals: number;
  assists: number;
  rating: number;
  captain: false;
  matches: number;
  yellowCards: number;
  redCards: number;
  minutesPlayed: number;
  shotsPerGame: number;
  passesPerGame: number;
  passAccuracy: number;
  dribbles: number;
  marketValue: string;
}